requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
Click==7.0
cligj==0.5.0
colorama==0.4.1
decorator==4.4.1
defusedxml==0.6.0
distro==1.4.0
dj-database-url==0.5.0
Django==2.2.5
django-bootstrap==0.2.4
django-crispy-forms==1.7.2
django-heroku==0.3.1
djangorestframework==3.11.0
entrypoints==0.3
Flask==1.1.1
geos==0.2.2
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
lxml==4.4.1
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.2.0
munch==2.5.0
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==4.4.0
nltk==3.4.5
notebook==6.0.1
numpy==1.17.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.8.3
Pygments==2.4.2
PyMySQL==0.9.3
PyPDF2==1.26.0
pyrsistent==0.15.5
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2019.2
pywin32==225
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.5
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.12.0
sqlparse==0.3.0
tabula-py==1.4.2
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
whitenoise==4.1.4
widgetsnbextension==3.5.1
zipp==0.6.0
